… buildconfig/Setup.Android.SDL2.in
… buildconfig/Setup.SDL2.in